Refactor Head into partial classes
